Use 64-bit integers for deserializing server responses
Users on 32-bit systems reported issues where the CLI would fail to
unmarshal responses from the server because the response contained
integer values that, when unmarshalled, would overflow 32-bit integers.
This commit changes all integers that we receive from the CC to
explicitly use 64 bit types.

We also changed the Uptime field of ccv3's ProcessInstance to a
time.Duration because it allowed us to isolate knowledge of
converting the uptime into a duration to the ProcessInstance's
UnmarshalJSON method.

[#163397517]

+++ actor/pushaction/command_line_settings.go
@@ -19,7 +19,7 @@ type CommandLineSettings struct {
 	DockerPassword       string
 	DockerUsername       string
 	DropletPath          string
-	HealthCheckTimeout   int
+	HealthCheckTimeout   uint64
 	HealthCheckType      string
 	Instances            types.NullInt
 	Memory               uint64
