+++ api/cloudcontroller/ccv2/application_instance_status.go
@@ -14,10 +14,10 @@ type ApplicationInstanceStatus struct {
 	CPU float64
 
 	// Disk is the instance's disk usage in bytes.
-	Disk int
+	Disk int64
 
 	// DiskQuota is the instance's allowed disk usage in bytes.
-	DiskQuota int
+	DiskQuota int64
 
 	// ID is the instance ID.
 	ID int
@@ -26,7 +26,7 @@ type ApplicationInstanceStatus struct {
 	IsolationSegment string
 
 	// Memory is the instance's memory usage in bytes.
-	Memory int
+	Memory int64
 
 	// MemoryQuota is the instance's allowed memory usage in bytes.
 	MemoryQuota int64
@@ -35,7 +35,7 @@ type ApplicationInstanceStatus struct {
 	State constant.ApplicationInstanceState
 
 	// Uptime is the number of seconds the instance has been running.
-	Uptime int
+	Uptime int64
 }
 
 // UnmarshalJSON helps unmarshal a Cloud Controller application instance
@@ -46,13 +46,13 @@ func (instance *ApplicationInstanceStatus) UnmarshalJSON(data []byte) error {
 		IsolationSegment string `json:"isolation_segment"`
 		Stats            struct {
 			Usage struct {
-				Disk   int     `json:"disk"`
-				Memory int     `json:"mem"`
+				Disk   int64   `json:"disk"`
+				Memory int64   `json:"mem"`
 				CPU    float64 `json:"cpu"`
 			} `json:"usage"`
 			MemoryQuota int64 `json:"mem_quota"`
-			DiskQuota   int   `json:"disk_quota"`
-			Uptime      int   `json:"uptime"`
+			DiskQuota   int64 `json:"disk_quota"`
+			Uptime      int64 `json:"uptime"`
 		} `json:"stats"`
 	}
 	err := cloudcontroller.DecodeJSON(data, &ccInstance)
